Thank you for your contribution to Spitball repo.
Before submitting this PR, please make sure:

- [ ] Your code builds clean without any errors
- [ ] You ran lint command on npm

+++ Cloudents.Web/Models/CreateCourseRequest.cs
@@ -17,10 +17,10 @@ namespace Cloudents.Web.Models
         [Required]
         public string Name { get; set; }
 
-        [Range(0,int.MaxValue)]
-        public int Price { get; set; }
+        [Range(0,double.MaxValue)]
+        public double Price { get; set; }
 
-        public int? SubscriptionPrice { get; set; }
+        public double? SubscriptionPrice { get; set; }
 
         [Required]
         public string Description { get; set; }
